+++ src/components/store/index.js
@@ -1,0 +1,3 @@
+export { default as StorePositionMap } from './StorePositionMap';
+export { default as Title } from './Title';
+export { default as DetailSide } from './DetailSide';
