+++ src/constants/index.js
@@ -4,3 +4,6 @@ export { default as categoryInfo } from './categoryInfo';
 export { default as storesQueryKey } from './storesQueryKey';
 export { default as searchedStoresQueryKey } from './searchedStoresQueryKey';
 export { default as STORES_FETCH_SIZE } from './storesFetchSize';
+export { default as commentQueryKey } from './commentQueryKey';
+export { default as storeQueryKey } from './storeQueryKey';
+export { default as archiveQueryKey } from './archiveQueryKey';
